+++ SQLiteStudio3/coreSQLiteStudio/parser/compile_lemon.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+LEMON=~/lemon
+GCC=gcc
+
+$GCC -o $LEMON lemon.c
+